Move _rpng_template functions to internal module
