+++ src/tqec/templates/library/hadamard.py
@@ -9,7 +9,6 @@ from tqec.templates.indices.qubit import (
     QubitTemplate,
     QubitVerticalBorders,
 )
-from tqec.templates.rpng import RPNGTemplate
 
 
 def get_temporal_hadamard_raw_template() -> QubitTemplate:
@@ -76,34 +75,6 @@ def get_temporal_hadamard_rpng_descriptions(
             13: RPNGDescription.from_string(f"-{bv}1h -{bv}2h ---- ----"),
         },
         default_factory=lambda: RPNGDescription.from_string("---- ---- ---- ----"),
-    )
-
-
-def get_temporal_hadamard_rpng_template(
-    orientation: ZObservableOrientation = ZObservableOrientation.HORIZONTAL,
-) -> RPNGTemplate[QubitTemplate]:
-    """Returns a representation of a transversal Hadamard gate applied on one
-    logical qubit.
-
-    This function returns a RPNGTemplate instance that performs a transversal
-    Hadamard gate. It is basically performing a standard memory operation, but
-    also include a Hadamard gate on each data-qubit, effectively performing a
-    fault-tolerant Hadamard gate by transversal application of physical gates.
-
-    Arguments:
-        orientation: orientation of the Z observable at the beginning of the
-            generated circuit description. The Z observable orientation will be
-            flipped at the end of the returned circuit description, which is
-            exactly the expected behaviour for a Hadamard transition.
-            Used to compute the stabilizers that should be measured on the
-            boundaries and in the bulk of the returned logical qubit description.
-
-    Returns:
-        an implementation of transversal Hadamard gate.
-    """
-    return RPNGTemplate(
-        template=get_temporal_hadamard_raw_template(),
-        mapping=get_temporal_hadamard_rpng_descriptions(orientation),
     )
 
 
@@ -190,44 +161,6 @@ def get_spatial_horizontal_hadamard_rpng_descriptions(
     )
 
 
-def get_spatial_horizontal_hadamard_rpng_template(
-    top_left_is_z_stabilizer: bool,
-    reset: Basis | None = None,
-    measurement: Basis | None = None,
-) -> RPNGTemplate[QubitHorizontalBorders]:
-    """Returns a representation of a "transversal" Hadamard gate at the interface
-    between two logical qubits aligned on the Y-axis
-
-    Note:
-        by convention, the hadamard-like transition is performed at the top-most
-        plaquettes.
-
-    Arguments:
-        top_left_is_z_stabilizer: if ``True``, the plaquette with index 5 in
-            :class:`~tqec.templates.indices.qubit.QubitHorizontalBorders`
-            should be measuring a Z stabilizer on its 2 top-most data-qubits and
-            a X stabilizer on its 2 bottom-most data-qubits. Else, it measures a
-            X stabilizer on its two top-most data-qubits and a Z stabilizer on
-            its two bottom-most data-qubits.
-        reset: basis of the reset operation performed on **internal**
-            data-qubits. Defaults to ``None`` that translates to no reset being
-            applied on data-qubits.
-        measurement: basis of the measurement operation performed on **internal**
-            data-qubits. Defaults to ``None`` that translates to no measurement
-            being applied on data-qubits.
-
-    Returns:
-        a description of a junction in the Y-axis and performing a logical
-        Hadamard operation.
-    """
-    return RPNGTemplate(
-        template=get_spatial_horizontal_hadamard_raw_template(),
-        mapping=get_spatial_horizontal_hadamard_rpng_descriptions(
-            top_left_is_z_stabilizer, reset, measurement
-        ),
-    )
-
-
 def get_spatial_vertical_hadamard_raw_template() -> QubitVerticalBorders:
     """Returns a representation of a "transversal" Hadamard gate at the interface
     between two logical qubits aligned on the X-axis.
@@ -308,42 +241,4 @@ def get_spatial_vertical_hadamard_rpng_descriptions(
             8: RPNGDescription.from_string(f"{r}{b2}1{m} -{b2}2- {r}{b2}3{m} -{b2}4-"),
         },
         default_factory=lambda: RPNGDescription.from_string("---- ---- ---- ----"),
-    )
-
-
-def get_spatial_vertical_hadamard_rpng_template(
-    top_left_is_z_stabilizer: bool,
-    reset: Basis | None = None,
-    measurement: Basis | None = None,
-) -> RPNGTemplate[QubitVerticalBorders]:
-    """Returns a representation of a "transversal" Hadamard gate at the interface
-    between two logical qubits aligned on the X-axis.
-
-    Note:
-        by convention, the hadamard-like transition is performed at the left-most
-        plaquettes.
-
-    Arguments:
-        top_left_is_z_stabilizer: if ``True``, the plaquette with index 5 in
-            :class:`~tqec.templates.indices.qubit.QubitVerticalBorders`
-            should be measuring a Z stabilizer on its 2 left-most data-qubits
-            and a X stabilizer on its 2 right-most data-qubits. Else, it
-            measures a X stabilizer on its two left-most data-qubits and a Z
-            stabilizer on its two right-most data-qubits.
-        reset: basis of the reset operation performed on **internal**
-            data-qubits. Defaults to ``None`` that translates to no reset being
-            applied on data-qubits.
-        measurement: basis of the measurement operation performed on **internal**
-            data-qubits. Defaults to ``None`` that translates to no measurement
-            being applied on data-qubits.
-
-    Returns:
-        a description of a junction in the X-axis and performing a logical
-        Hadamard operation.
-    """
-    return RPNGTemplate(
-        template=get_spatial_vertical_hadamard_raw_template(),
-        mapping=get_spatial_vertical_hadamard_rpng_descriptions(
-            top_left_is_z_stabilizer, reset, measurement
-        ),
     )
