+++ src/tqec/templates/library/spatial.py
@@ -25,7 +25,6 @@ from tqec.templates.indices.qubit import (
     QubitSpatialJunctionTemplate,
     QubitVerticalBorders,
 )
-from tqec.templates.rpng import RPNGTemplate
 
 
 def get_spatial_junction_qubit_raw_template() -> QubitSpatialJunctionTemplate:
@@ -279,72 +278,6 @@ def get_spatial_junction_qubit_rpng_descriptions(
     )
 
 
-def get_spatial_junction_qubit_rpng_template(
-    spatial_boundary_basis: Basis,
-    arms: JunctionArms,
-    reset: Basis | None = None,
-    measurement: Basis | None = None,
-) -> RPNGTemplate[QubitSpatialJunctionTemplate]:
-    """Implementation of a logical qubit acting as a spatial junction.
-
-    This function returns a RPNGTemplate instance representing a logical qubit
-    connecting to others in space and forming a spatial junction. The returned
-    template is carefully crafted to avoid hook errors damaging the logical
-    distance.
-
-    Note:
-        this function does not enforce anything on the input values. As such, it
-        is possible to generate a description of a round that will both reset and
-        measure the data-qubits.
-
-    Warning:
-        By convention, this function does not populate the plaquettes on the
-        boundaries where an arm is present **BUT** do populate the corners (that
-        are part of the boundaries, so this is an exception to the first part of
-        the sentence).
-
-        The rationale behind that convention is that the logical qubit
-        representing the spatial junction is completely aware of all the arms
-        that should be implemented whereas each arm in isolation does not know
-        if the other arms are present or not. That means that corners, whose
-        plaquette depends on the presence or absence of the two arms it belongs
-        to, require information that is given to this function, but not to the
-        arm-generation function.
-
-        Arms should follow that convention and should not replace the plaquette
-        descriptions on the corners (i.e., not include an explicit mapping, even
-        to the empty plaquette, from the index of the corner to a plaquette).
-
-
-    Arguments:
-        spatial_boundary_basis: stabilizers that are measured at each boundaries
-            of the spatial junction.
-        arms: flag-like enumeration listing the arms that are used around the
-            logical qubit. The returned template will be adapted to be
-            compatible with such a layout.
-        reset: basis of the reset operation performed on data-qubits. Defaults
-            to ``None`` that translates to no reset being applied on data-qubits.
-        measurement: basis of the measurement operation performed on data-qubits.
-            Defaults to ``None`` that translates to no measurement being applied
-            on data-qubits.
-
-    Raises:
-        TQECException: if ``arms`` only contains 0 or 1 flag.
-        TQECException: if ``arms`` describes an I-shaped junction (TOP/DOWN or
-            LEFT/RIGHT).
-
-    Returns:
-        a description of a logical qubit performing a memory operation while
-        being enclosed by 2 or more arms.
-    """
-    return RPNGTemplate(
-        template=get_spatial_junction_qubit_raw_template(),
-        mapping=get_spatial_junction_qubit_rpng_descriptions(
-            spatial_boundary_basis, arms, reset, measurement
-        ),
-    )
-
-
 def get_spatial_junction_arm_raw_template(
     arm: JunctionArms,
 ) -> QubitVerticalBorders | QubitHorizontalBorders:
@@ -470,62 +403,6 @@ def get_spatial_junction_arm_rpng_descriptions(
             raise TQECException(
                 f"The 'arm' parameter should contain exactly 1 flag. Got {arm}."
             )
-
-
-def get_spatial_junction_arm_template(
-    spatial_boundary_basis: Basis,
-    arm: JunctionArms,
-    reset: Basis | None = None,
-    measurement: Basis | None = None,
-) -> RPNGTemplate[QubitVerticalBorders | QubitHorizontalBorders]:
-    """Implementation of arms for a spatial junction around a logical qubit.
-
-    This function returns a RPNGTemplate instance representing the arms
-    required to perform a spatial junction on a logical qubit that has 2 or more
-    arms. The returned template is carefully crafted to avoid hook errors
-    damaging the logical distance.
-
-    Note:
-        this function does not enforce anything on the input values. As such, it
-        is possible to generate a description of a round that will both reset and
-        measure the data-qubits.
-
-    Warning:
-        by convention, this function should **not** populate the plaquettes on
-        the corners as :func:`get_spatial_junction_qubit_template` should take
-        care of that.
-
-    Warning:
-        Using this function without :func:`get_spatial_junction_qubit_template`
-        is very likely a programming error. Please double-check what you are
-        doing if that is your case, in particular how the plaquettes on each
-        corner of the center logical qubit are set.
-
-    Arguments:
-        spatial_boundary_basis: stabilizers that are measured at each boundaries
-            of the spatial junction.
-        arm: arm to return a spatial junction for. Should contain exactly
-            **one** of the possible arm flags.
-        reset: basis of the reset operation performed on **internal**
-            data-qubits. Defaults to ``None`` that translates to no reset being
-            applied on data-qubits.
-        measurement: basis of the measurement operation performed on **internal**
-            data-qubits. Defaults to ``None`` that translates to no measurement
-            being applied on data-qubits.
-
-    Raises:
-        TQECException: if ``arm`` does not contain exactly 1 flag (i.e., if it
-            contains 0 or 2+ flags).
-
-    Returns:
-        a description of the provided ``arm``.
-    """
-    return RPNGTemplate(
-        template=get_spatial_junction_arm_raw_template(arm),
-        mapping=get_spatial_junction_arm_rpng_descriptions(
-            spatial_boundary_basis, arm, reset, measurement
-        ),
-    )
 
 
 def _get_left_spatial_junction_arm_rpng_descriptions(
